sespro-fastapi/requirements.txt
aiofiles==0.6.0
appdirs==1.4.3
asgi-lifespan==1.0.1
attrs==20.3.0
bcrypt==3.2.0
boto3==1.17.32
botocore==1.20.32
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
Faker==6.6.2
fastapi==0.63.0
fastapi-versioning==0.10.0
google-auth==1.27.1
greenlet==1.0.0
h11==0.12.0
html5lib==1.0.1
httpcore==0.12.3
httpx==0.17.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
Jinja2==2.11.3
jmespath==0.10.0
jose==1.0.0
lockfile==0.12.2
mangum==0.11.0
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==10.0.0
pluggy==0.13.1
pony==0.7.14
progress==1.5
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pyparsing==2.4.6
pytest==6.2.2
pytest-asyncio==0.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytoml==0.1.21
PyYAML==5.4.1
pytz==2021.1
requests==2.22.0
retrying==1.3.3
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.3.6
six==1.14.0
sniffio==1.2.0
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.25.8
uvicorn==0.13.4
validators==0.18.2
watchtower==1.0.6
webencodings==0.5.1
yapf==0.30.0
